+++ tools/package.py
@@ -34,6 +34,7 @@ EXPECTED = [
     "Plugins/KRPC.Bridge.Fmrs.dll",
     "Plugins/KRPC.Bridge.Ocisly.dll",
     "Plugins/KRPC.Bridge.MechJeb.dll",
+    "Plugins/KRPC.Bridge.Trajectories.dll",
 ]
 
 EXTRAS = [